+++ src/main/java/app/App.java
@@ -1,4 +1,4 @@
-package org.example;
+package app;
 
 public class App {
     public static void main(String[] args) {
