+++ src/test/java/app/AppTest.java
@@ -1,12 +1,12 @@
-package org.example;
+package app;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class AppIT {
+class AppTest {
     @Test
-    void itTest() {
-        assertThat(false).isFalse();
+    void test() {
+        assertThat(true).isTrue();
     }
 }
